move lightnode page to root
