+++ containers/lightnode/howToJoinSection.tsx
