+++ containers/lightnode/howToRelayRewardsSection.tsx
