+++ containers/lightnode/howToWorkSection.tsx
