+++ containers/lightnode/investorSection.tsx
